+++ FinEdgeBackend/DTOs/Transactions/TransactionDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace FinEdgeBackend.DTOs
+﻿namespace FinEdgeBackend.DTOs.Transactions
 {
     public class TransactionDTO
     {
