+++ FinEdgeBackend/Services/TransactionService.cs
@@ -1,5 +1,5 @@
 ﻿using FinEdgeBackend.Data;
-using FinEdgeBackend.DTOs;
+using FinEdgeBackend.DTOs.Transactions;
 using FinEdgeBackend.Interfaces;
 using FinEdgeBackend.Models;
 using Microsoft.EntityFrameworkCore;
@@ -57,6 +57,63 @@ namespace FinEdgeBackend.Services
         public ICollection<Transaction> GetTransactionsFromSpecifiedDate(ICollection<Transaction> transactions, DateTime date)
         {
             return transactions.Where(t => t.DateCreated >= date).ToList();
+        }
+
+        public TransactionSankeyChartDTO GetSankeyChartDataAsync(ICollection<Category> categories)
+        {
+            ICollection<Category> incomeCategories = _categoryService.GetIncomeCategories(categories);
+            ICollection<Category> expenditureCategories = _categoryService.GetExpenditureCategories(categories);
+
+            decimal totalIncome = incomeCategories.Sum(c => c.Balance ?? 0);
+            decimal totalExpenditure = expenditureCategories.Sum(c => c.Balance ?? 0);
+
+            TransactionSankeyChartDTO sankeyChartDto = new TransactionSankeyChartDTO()
+            {
+                Nodes = new List<SankeyChartNodeDTO>
+                {
+                    new SankeyChartNodeDTO()
+                    {
+                        Id = "income",
+                        Label = new SankeyChartLabelDTO
+                        {
+                            Text = $"Income ({totalIncome:C})"
+                        }
+                    },
+                    new SankeyChartNodeDTO()
+                    {
+                        Id = "expenditure",
+                        Label = new SankeyChartLabelDTO
+                        {
+                            Text = $"Expenditure ({totalExpenditure:C})"
+                        }
+                    }
+                },
+                Links = new List<SankeyChartLinkDTO>()
+            };
+
+            foreach (Category category in categories)
+            {
+                sankeyChartDto.Nodes.Add(new SankeyChartNodeDTO()
+                {
+                    Id = category.Name!.ToLower().Replace(" ", "_"),
+                    Label = new SankeyChartLabelDTO
+                    {
+                        Text = $"{category.Name} ({category.Balance:C})"
+                    }
+                });
+            }
+
+            foreach (Category category in categories)
+            {
+                sankeyChartDto.Links.Add(new SankeyChartLinkDTO()
+                {
+                    SourceId = category.IsIncome ? "income" : "expenditure",
+                    TargetId = category.Name!.ToLower().Replace(" ", "_"),
+                    Value = category.Balance ?? 0
+                });
+            }
+
+            return sankeyChartDto;
         }
 
         public async Task UpdateTranssactionAsync(TransactionDTO transactionDto, Transaction transaction, Category category, Account newAccount, Account originalAccount, User currentUser)
